Better partial feedback for word validity (fix #10)
This seems more satisfactory. We show when words *aren't* valid,
and in such a way that crossing information is still visibile.

Might still want to consider changing the concrete appearance of the
annotations, specifically as to what side they appear on (maybe
both?) or what color they are.

Note that this means the computation of *valid* words is presently
dead code now. Will leave it in for now in case I need it later.