Fix rendering problem.
To reproduce: select a bunch of tiles, zoom in or out, and then try to
drag them to the hand. The spacing between tiles is wrong. The
solution was to compose a world-to-world transform (or, well, if I
were more careful, I could construe it as a undragged-tile-frame to
undragged-tile-frame transform) from the other end.