Refactor layer handling (fix #26)
Instead of making it less global, I made it more global.
A "Layer" is really a functional generator for point-indexed data.
The cache is an impure performance optimization of a pure function.
Therefore it's acceptable to make it global (as long as I prevent
multiple such caches from interfering with one another, which I'm
now doing by requiring a unique name to be passed in)