Alter active chunk calculation slightly
This should equivalently ensure that any point within
`world_bds_in_canvas` can be rendered, but makes more sense to me
logically. If we have a point p in world coordinates, the chunk that
"owns" it is always ⌊p⌋. It should therefore suffice to do an
end-point inclusive for-loop from the min required chunk to the max
required chunk.

This should adapt more sensibly to the near-future world where I want
to add on some extra margin for neighbor sampling.