[GL] Remove deprecated functional cache maintenance strategy
It makes me a little sad to do this, but at least this moves me to a
state where the code isn't trying to ambivalently maintain the cache
state in two different ways.

The two general approaches to cache-during-drag are:

1) Make a copy of the cache and store it with the mouse state during
drag. Have GL rendering use this rather than the main state when
mouseState is drag.

2) Statefully update the cache when drag begins and ends.

I was starting out trying to do (1), and gradually made some changes
and bugfixes that amounted to doing (2), and now I've fully committed
to (2) for now. Going back to (1) would mean being able to eliminate
the `restoreTile` lowAction, I think?